+++ sdk/src/client/secret/placeholder.rs
@@ -30,7 +30,7 @@ impl SecretManage for PlaceholderSecretManager {
         _coin_type: u32,
         _account_index: u32,
         _address_indexes: Range<u32>,
-        _options: Option<GenerateAddressOptions>,
+        _options: impl Into<Option<GenerateAddressOptions>> + Send,
     ) -> Result<Vec<Ed25519Address>, Self::Error> {
         Err(Error::PlaceholderSecretManager)
     }
@@ -40,7 +40,7 @@ impl SecretManage for PlaceholderSecretManager {
         _coin_type: u32,
         _account_index: u32,
         _address_indexes: Range<u32>,
-        _options: Option<GenerateAddressOptions>,
+        _options: impl Into<Option<GenerateAddressOptions>> + Send,
     ) -> Result<Vec<EvmAddress>, Self::Error> {
         Err(Error::PlaceholderSecretManager)
     }