+++ sdk/src/client/stronghold/secret.rs
@@ -38,7 +38,7 @@ impl SecretManage for StrongholdAdapter {
         coin_type: u32,
         account_index: u32,
         address_indexes: Range<u32>,
-        options: Option<GenerateAddressOptions>,
+        options: impl Into<Option<GenerateAddressOptions>> + Send,
     ) -> Result<Vec<Ed25519Address>, Self::Error> {
         // Prevent the method from being invoked when the key has been cleared from the memory. Do note that Stronghold
         // only asks for a key for reading / writing a snapshot, so without our cached key this method is invocable, but
@@ -54,7 +54,7 @@ impl SecretManage for StrongholdAdapter {
 
         // Addresses to return.
         let mut addresses = Vec::new();
-        let internal = options.map(|o| o.internal).unwrap_or_default();
+        let internal = options.into().map(|o| o.internal).unwrap_or_default();
 
         for address_index in address_indexes {
             let bip_path = [HD_WALLET_TYPE, coin_type, account_index, internal as u32, address_index];
@@ -102,7 +102,7 @@ impl SecretManage for StrongholdAdapter {
         coin_type: u32,
         account_index: u32,
         address_indexes: Range<u32>,
-        options: Option<GenerateAddressOptions>,
+        options: impl Into<Option<GenerateAddressOptions>> + Send,
     ) -> Result<Vec<EvmAddress>, Self::Error> {
         // Prevent the method from being invoked when the key has been cleared from the memory. Do note that Stronghold
         // only asks for a key for reading / writing a snapshot, so without our cached key this method is invocable, but
@@ -118,7 +118,7 @@ impl SecretManage for StrongholdAdapter {
 
         // Addresses to return.
         let mut addresses = Vec::new();
-        let internal = options.map(|o| o.internal).unwrap_or_default();
+        let internal = options.into().map(|o| o.internal).unwrap_or_default();
 
         for address_index in address_indexes {
             let chain = Chain::from_u32_hardened([HD_WALLET_TYPE, coin_type, account_index])