+++ sdk/tests/wallet/wallet.rs
@@ -185,7 +185,7 @@ async fn wallet_address_generation() -> Result<()> {
     }
     let wallet = wallet_builder.finish().await?;
 
-    let address = wallet.generate_address(0, 0, None).await?;
+    let address = wallet.generate_ed25519_address(0, 0, None).await?;
 
     assert_eq!(
         address.to_bech32_unchecked("smr"),
@@ -214,7 +214,7 @@ async fn wallet_address_generation() -> Result<()> {
         }
         let wallet = wallet_builder.finish().await?;
 
-        let address = wallet.generate_address(0, 0, None).await?;
+        let address = wallet.generate_ed25519_address(0, 0, None).await?;
 
         assert_eq!(
             address.to_bech32_unchecked("smr"),